remove auditRound.measuredMean bump version 1.7.2
